Refactor RawSensorInfo
